neuroglancer: Renamed annotations.py -> annotations/local.py
