+++ neuclease/misc/neuroglancer/__init__.py
@@ -5,6 +5,6 @@ See also: neuclease/notebooks/hemibrain-neuroglancer-video-script.txt
 """
 from .storage import download_ngstate, upload_ngstate, upload_ngstates, upload_json, upload_to_bucket, make_bucket_public
 from .util import parse_nglink, format_nglink, layer_dict, layer_state
-from .annotations import extract_annotations, annotation_layer_json, point_annotation_layer_json
+from .annotations.local import extract_annotations, annotation_layer_json, point_annotation_layer_json
 from .segmentprops import segment_properties_json
 from .segmentcolors import hex_string_from_segment_id
